+++ JoeySoft/JoeySoft.TfsDevelop/JoeySoft.TfsDevelop/JoeySoft.TfsDevelop.csproj
@@ -59,9 +59,9 @@
       <Project>{5d55eca0-96e1-4696-8abf-7fc7da42350f}</Project>
       <Name>JoeySoft.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\JoeySofy.TFS\JoeySofy.TFS.csproj">
+    <ProjectReference Include="..\JoeySofy.TFS\JoeySoft.TFS.csproj">
       <Project>{dfdee644-0d25-4355-86dc-230458c7e034}</Project>
-      <Name>JoeySofy.TFS</Name>
+      <Name>JoeySoft.TFS</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
