+++ JoeySoft/JoeySoft.TfsDevelop/JoeySoft.TfsDevelopWinFrom/TfsDevelopFrom.resx
@@ -125,7 +125,7 @@
         AAEAAAD/////AQAAAAAAAAAMAgAAAFdTeXN0ZW0uV2luZG93cy5Gb3JtcywgVmVyc2lvbj00LjAuMC4w
         LCBDdWx0dXJlPW5ldXRyYWwsIFB1YmxpY0tleVRva2VuPWI3N2E1YzU2MTkzNGUwODkFAQAAACZTeXN0
         ZW0uV2luZG93cy5Gb3Jtcy5JbWFnZUxpc3RTdHJlYW1lcgEAAAAERGF0YQcCAgAAAAkDAAAADwMAAACo
-        DgAAAk1TRnQBSQFMAgEBBgEAATABAAEwAQABEAEAARABAAT/ARkBAAj/AUIBTQE2BwABNgMAASgDAAFA
+        DgAAAk1TRnQBSQFMAgEBBgEAATgBAAE4AQABEAEAARABAAT/ARkBAAj/AUIBTQE2BwABNgMAASgDAAFA
         AwABIAMAAQEBAAEYBgABGBIAYP9gAGD/YAAD/wEAAYACAAGAAgABgAIAAYACAAGAAgABgAIAAYACAAGA
         AgABgAIAAYACAAGAAgABgAIAAYABAAn/AQABgAIAAYACAAGAAgABgAIAAYACAAGAAgABgAIAAYACAAGA
         AgABgAIAAYACAAGAAgABgAEABv9gAAP/AQABgAEAIf8BAAGAAQAJ/wEAAYABACH/AQABgAEABv9gAAP/
