+++ JoeySoft/JoeySoft.TfsDevelop/JoeySofy.TFS/JoeySoft.TFS.csproj
